+++ quadratic-core/src/controller.rs
@@ -1,12 +1,11 @@
-use serde::{Deserialize, Serialize};
 #[cfg(feature = "js")]
 use wasm_bindgen::prelude::*;
 
-use crate::{grid::*, Array, CellValue, Pos, Rect, RunLengthEncoding};
+use crate::grid::Grid;
 
-use self::transactions::{Operation, Transaction, TransactionSummary};
+use self::transactions::Transaction;
 
-mod controller_tests;
+pub mod cells;
 pub mod sheets;
 pub mod transactions;
 
@@ -28,154 +27,4 @@ impl GridController {
             redo_stack: vec![],
         }
     }
-    pub fn populate_with_random_floats(&mut self, sheet_id: SheetId, region: &Rect) {
-        let sheet = self.grid.sheet_mut_from_id(sheet_id);
-        sheet.with_random_floats(region);
-    }
-
-    pub fn grid(&self) -> &Grid {
-        &self.grid
-    }
-
-    pub fn set_cell_value(
-        &mut self,
-        sheet_id: SheetId,
-        pos: Pos,
-        value: CellValue,
-        cursor: Option<String>,
-    ) -> TransactionSummary {
-        let sheet = self.grid.sheet_mut_from_id(sheet_id);
-        let cell_ref = sheet.get_or_create_cell_ref(pos);
-        let region = RegionRef::from(cell_ref);
-        let values = Array::from(value);
-        let ops = vec![Operation::SetCellValues { region, values }];
-        self.transact_forward(Transaction { ops, cursor })
-    }
-    pub fn set_cells(
-        &mut self,
-        sheet_id: SheetId,
-        start_pos: Pos,
-        values: Array,
-        cursor: Option<String>,
-    ) -> TransactionSummary {
-        let end_pos = Pos {
-            x: start_pos.x + values.width() as i64 - 1,
-            y: start_pos.y + values.height() as i64 - 1,
-        };
-        let rect = Rect {
-            min: start_pos,
-            max: end_pos,
-        };
-        let region = self.region(sheet_id, rect);
-        let ops = vec![Operation::SetCellValues { region, values }];
-        self.transact_forward(Transaction { ops, cursor })
-    }
-    pub fn delete_cell_values(
-        &mut self,
-        sheet_id: SheetId,
-        rect: Rect,
-        cursor: Option<String>,
-    ) -> TransactionSummary {
-        let region = self.existing_region(sheet_id, rect);
-        let ops = match region.size() {
-            Some(size) => {
-                let values = Array::new_empty(size);
-                vec![Operation::SetCellValues { region, values }]
-            }
-            None => vec![], // region is empty; do nothing
-        };
-        self.transact_forward(Transaction { ops, cursor })
-    }
-
-    fn set_cell_formats_for_type<A: CellFmtAttr>(
-        &mut self,
-        region: &RegionRef,
-        values: RunLengthEncoding<Option<A::Value>>,
-    ) -> RunLengthEncoding<Option<A::Value>> {
-        let sheet = self.grid.sheet_mut_from_id(region.sheet);
-        // TODO: optimize this for contiguous runs of the same value
-        let mut old_values = RunLengthEncoding::new();
-        for (cell_ref, value) in region.iter().zip(values.iter_values()) {
-            let old_value = sheet
-                .cell_ref_to_pos(cell_ref)
-                .and_then(|pos| sheet.set_formatting_value::<A>(pos, value.clone()));
-            old_values.push(old_value);
-        }
-        old_values
-    }
-
-    /// Returns a region of the spreadsheet, assigning IDs to columns and rows
-    /// as needed.
-    fn region(&mut self, sheet_id: SheetId, rect: Rect) -> RegionRef {
-        let sheet = self.grid.sheet_mut_from_id(sheet_id);
-        let columns = rect
-            .x_range()
-            .map(|x| sheet.get_or_create_column(x).0.id)
-            .collect();
-        let rows = rect
-            .y_range()
-            .map(|y| sheet.get_or_create_row(y).id)
-            .collect();
-        RegionRef {
-            sheet: sheet_id,
-            columns,
-            rows,
-        }
-    }
-    /// Returns a region of the spreadsheet, ignoring columns and rows which
-    /// have no contents and no IDs.
-    fn existing_region(&self, sheet_id: SheetId, rect: Rect) -> RegionRef {
-        let sheet = self.grid.sheet_from_id(sheet_id);
-        let columns = rect
-            .x_range()
-            .filter_map(|x| sheet.get_column(x))
-            .map(|col| col.id)
-            .collect();
-        let rows = rect.y_range().filter_map(|y| sheet.get_row(y)).collect();
-        RegionRef {
-            sheet: sheet_id,
-            columns,
-            rows,
-        }
-    }
-}
-
-macro_rules! impl_set_cell_fmt_method {
-    ($method_name:ident<$cell_fmt_attr_type:ty>($cell_fmt_array_constructor:expr)) => {
-        impl GridController {
-            pub fn $method_name(
-                &mut self,
-                sheet_id: SheetId,
-                rect: Rect,
-                value: Option<<$cell_fmt_attr_type as CellFmtAttr>::Value>,
-                cursor: Option<String>,
-            ) -> TransactionSummary {
-                let region = self.region(sheet_id, rect);
-                let attr =
-                    $cell_fmt_array_constructor(RunLengthEncoding::repeat(value, region.len()));
-                let ops = vec![Operation::SetCellFormats { region, attr }];
-                self.transact_forward(Transaction { ops, cursor })
-            }
-        }
-    };
-}
-
-impl_set_cell_fmt_method!(set_cell_align<CellAlign>(CellFmtArray::Align));
-impl_set_cell_fmt_method!(set_cell_wrap<CellWrap>(CellFmtArray::Wrap));
-impl_set_cell_fmt_method!(set_cell_numeric_format<NumericFormat>(CellFmtArray::NumericFormat));
-impl_set_cell_fmt_method!(set_cell_bold<Bold>(CellFmtArray::Bold));
-impl_set_cell_fmt_method!(set_cell_italic<Italic>(CellFmtArray::Italic));
-impl_set_cell_fmt_method!(set_cell_text_color<TextColor>(CellFmtArray::TextColor));
-impl_set_cell_fmt_method!(set_cell_fill_color<FillColor>(CellFmtArray::FillColor));
-
-/// Array of a single cell formatting attribute.
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub enum CellFmtArray {
-    Align(RunLengthEncoding<Option<CellAlign>>),
-    Wrap(RunLengthEncoding<Option<CellWrap>>),
-    NumericFormat(RunLengthEncoding<Option<NumericFormat>>),
-    Bold(RunLengthEncoding<Option<bool>>),
-    Italic(RunLengthEncoding<Option<bool>>),
-    TextColor(RunLengthEncoding<Option<String>>),
-    FillColor(RunLengthEncoding<Option<String>>),
 }
