+++ quadratic-core/src/controller/sheets.rs
@@ -113,3 +113,79 @@ impl GridController {
         self.transact_forward(Transaction { ops, cursor })
     }
 }
+
+#[test]
+fn test_add_delete_reorder_sheets() {
+    let mut g = GridController::new();
+    g.add_sheet(None);
+    g.add_sheet(None);
+    let old_sheet_ids = g.sheet_ids();
+    let s1 = old_sheet_ids[0];
+    let s2 = old_sheet_ids[1];
+    let s3 = old_sheet_ids[2];
+
+    let mut test_reorder = |a, b, expected: [SheetId; 3]| {
+        g.move_sheet(a, b, None);
+        assert_eq!(expected.to_vec(), g.sheet_ids());
+        g.undo(None);
+        assert_eq!(old_sheet_ids, g.sheet_ids());
+    };
+
+    test_reorder(s1, Some(s2), [s1, s2, s3]);
+    test_reorder(s1, Some(s3), [s2, s1, s3]);
+    test_reorder(s1, None, [s2, s3, s1]);
+    test_reorder(s2, Some(s1), [s2, s1, s3]);
+    test_reorder(s2, Some(s3), [s1, s2, s3]);
+    test_reorder(s2, None, [s1, s3, s2]);
+    test_reorder(s3, Some(s1), [s3, s1, s2]);
+    test_reorder(s3, Some(s2), [s1, s3, s2]);
+    test_reorder(s3, None, [s1, s2, s3]);
+
+    let mut test_delete = |a, expected: [SheetId; 2]| {
+        g.delete_sheet(a, None);
+        assert_eq!(expected.to_vec(), g.sheet_ids());
+        g.undo(None);
+        assert_eq!(old_sheet_ids, g.sheet_ids());
+    };
+
+    test_delete(s1, [s2, s3]);
+    test_delete(s2, [s1, s3]);
+    test_delete(s3, [s1, s2]);
+}
+
+#[test]
+fn test_duplicate_sheet() {
+    let mut g = GridController::new();
+    let old_sheet_ids = g.sheet_ids();
+    let s1 = old_sheet_ids[0];
+
+    g.set_sheet_name(s1, String::from("Nice Name"), None);
+    g.duplicate_sheet(s1, None);
+    let sheet_ids = g.sheet_ids();
+    let s2 = sheet_ids[1];
+
+    let sheet1 = g.sheet(s1);
+    let sheet2 = g.sheet(s2);
+
+    assert_eq!(sheet2.name, format!("{} Copy", sheet1.name));
+}
+
+#[test]
+fn test_delete_last_sheet() {
+    let mut g = GridController::new();
+    let sheet_ids = g.sheet_ids();
+    let first_sheet_id = sheet_ids[0].clone();
+
+    g.delete_sheet(first_sheet_id, None);
+    let new_sheet_ids = g.sheet_ids();
+    assert_eq!(new_sheet_ids.len(), 1);
+    assert_ne!(new_sheet_ids[0], sheet_ids[0]);
+
+    g.undo(None);
+    let new_sheet_ids_2 = g.sheet_ids();
+    assert_eq!(sheet_ids[0], new_sheet_ids_2[0]);
+
+    g.redo(None);
+    let new_sheet_ids_3 = g.sheet_ids();
+    assert_eq!(new_sheet_ids[0], new_sheet_ids_3[0]);
+}
